requirements.txt
# Python dependencies - Compatible with Python 3.13
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles>=23.0.0

# ML/AI Libraries - Using flexible versions for compatibility
transformers>=4.40.0
sentence-transformers>=2.7.0
torch>=2.6.0
numpy
pandas

# Google AI Integration
google-generativeai>=0.3.0

# OCR & Document Processing
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2==3.0.1
pdf2image>=1.16.0

# NLP & NER - Flexible versions
spacy>=3.6.0
nltk>=3.8.0

# Search & Vector Database
faiss-cpu>=1.7.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0

# HTTP Client & API
httpx>=0.25.0
requests>=2.31.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-slugify>=8.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx==0.27.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
